requirements.txt
Django==5.1.2
python-decouple==3.8
django-ninja==1.3.0